requirements.txt
flask
requests
opencv-python-headless
pillow
numpy
